Extract into dedicated class and add compilation error test
