+++ src/NServiceBus.Core.Analyzer.Tests.Roslyn5/ApprovalFiles/AddHandlerGeneratorCollisionTests.HandlerNameCollisionResolvedByRegistrationMethodNamePatterns.approved.txt
